requirements.txt
Django==1.9.1
django-allauth==0.24.1
django-extensions==1.6.1
djangorestframework==3.3.2
djangorestframework-jsonp==1.0.2
oauthlib==1.0.3
psycopg2==2.6.1
python-openid==2.2.5
pytz==2015.7
PyYAML==3.11
requests==2.9.1
requests-oauthlib==0.6.0
urllib3==1.14
Pillow==4.2.1
